sql/schemachanger/sctest: pull end-to-end tests to a runner package
This will help to write CCL tests.

Release note: None
