+++ pkg/sql/schemachanger/BUILD.bazel
@@ -21,16 +21,10 @@ go_test(
         "//pkg/sql",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/desctestutils",
-        "//pkg/sql/parser",
-        "//pkg/sql/schemachanger/scbuild",
-        "//pkg/sql/schemachanger/scdeps/sctestdeps",
-        "//pkg/sql/schemachanger/scdeps/sctestutils",
         "//pkg/sql/schemachanger/scop",
-        "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
         "//pkg/sql/schemachanger/scrun",
-        "//pkg/sql/sessiondata",
-        "//pkg/sql/sessiondatapb",
+        "//pkg/sql/schemachanger/sctest",
         "//pkg/sql/tests",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
@@ -40,8 +34,6 @@ go_test(
         "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/randutil",
-        "@com_github_cockroachdb_cockroach_go_v2//crdb",
-        "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
