+++ pkg/sql/schemachanger/sctest/BUILD.bazel
@@ -1,0 +1,28 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "sctest",
+    srcs = ["end_to_end.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/schemachanger/sctest",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/jobs/jobspb",
+        "//pkg/keys",
+        "//pkg/sql/parser",
+        "//pkg/sql/schemachanger/scbuild",
+        "//pkg/sql/schemachanger/scdeps/sctestdeps",
+        "//pkg/sql/schemachanger/scdeps/sctestutils",
+        "//pkg/sql/schemachanger/scop",
+        "//pkg/sql/schemachanger/scpb",
+        "//pkg/sql/schemachanger/scplan",
+        "//pkg/sql/schemachanger/scrun",
+        "//pkg/sql/sessiondata",
+        "//pkg/sql/sessiondatapb",
+        "//pkg/testutils",
+        "//pkg/testutils/sqlutils",
+        "@com_github_cockroachdb_cockroach_go_v2//crdb",
+        "@com_github_cockroachdb_datadriven//:datadriven",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
+    ],
+)
